moved scratch space related code for PQ flash index to scratch.h/cpp
