+++ include/scratch.h
@@ -13,6 +13,13 @@
 
 #include "neighbor.h"
 #include "concurrent_queue.h"
+#include "pq.h"
+#include "aligned_file_reader.h"
+
+#define MAX_GRAPH_DEGREE 512
+#define MAX_N_CMPS 16384
+#define SECTOR_LEN (_u64) 4096
+#define MAX_N_SECTOR_READS 128
 
 namespace diskann {
   template<typename T>
@@ -56,6 +63,37 @@ namespace diskann {
     boost::dynamic_bitset<> &inserted_into_pool_bs() {
       return *_inserted_into_pool_bs;
     }
+  };
+
+  template<typename T>
+  struct SSDQueryScratch {
+    T   *coord_scratch = nullptr;  // MUST BE AT LEAST [MAX_N_CMPS * data_dim]
+    _u64 coord_idx = 0;            // index of next [data_dim] scratch to use
+
+    char *sector_scratch =
+        nullptr;          // MUST BE AT LEAST [MAX_N_SECTOR_READS * SECTOR_LEN]
+    _u64 sector_idx = 0;  // index of next [SECTOR_LEN] scratch to use
+
+    T *aligned_query_T = nullptr;
+
+    PQScratch<T> *_pq_scratch;
+
+    tsl::robin_set<_u64>  visited;
+    std::vector<Neighbor> retset;
+    std::vector<Neighbor> full_retset;
+
+    SSDQueryScratch(size_t aligned_dim, size_t visited_reserve);
+    ~SSDQueryScratch();
+
+    void reset();
+  };
+
+  template<typename T>
+  struct SSDThreadData {
+    SSDQueryScratch<T> scratch;
+    IOContext          ctx;
+
+    SSDThreadData(size_t aligned_dim, size_t visited_reserve);
   };
 
   // Class to avoid the hassle of pushing and popping the query scratch.
